Fix crash when replacing instrument via popup in instruments panel
The problem is that the popup is unloaded as soon as it is closed, and it is closed as soon as the "replace instrument" dialog appears. The unloading causes the popup and its content and the model of that content to be deleted. But the "onClicked" signal handler is still being executed at that point. This is (naturally) not allowed by Qt and causes a crash.

Probably introduced by 95249f7043efb2d67008c4ec47f36aa105867948; that commit fixed a mistake, but that mistake prevented the unloading, and thus the crash.