Remove ledger lines when no longer needed
When going from e.g. 2 to 1 ledger line, the unneeded ledger line is deleted by the call to `resizeLedgerLinesTo` at the bottom of `updateLedgerLines`. But when going from any number to 0 ledger lines, there was an early return, and the ledger lines were not deleted.

Since they would not be laid out, and all element's bounding boxes are reset before the layout process, they would have bounding box zero; and thus, would not be seen by `Page::items`, so would not be drawn in the notation view. But SVG export uses Page::elements` instead (which makes sense), which does see elements with empty bounding boxes.